Chapter 1: Inception Practice Assignment
THEORY ASSIGNMENT - 1

1. What is Emmet?
Ans:- 
Emmet is a plugin for text editors that allow for high speed coding and editing in HTML,XML & gives auto generated code for the same.

2. Difference between a Library and Framework?
Ans:- 
1. By using library, you can control flow of application all call the library. By using framework, the control is inverted, i.e., the framework controls the flow and calls your code.
2. (Inversion control point of view) Libraries are a collection of helper modules, objects, classes, functions, message templates, pre-written code, etc. Whereas, Frameworks consist of a lot of APIs, compilers, toolsets, support programs, libraries, etc.
3. In performance point of view, less code is required to build libraries, which leads to faster loading times and better performance. Whereas, developing a framework requires a lot of coding, which increases loading times and decreases performance.
4. In existing project, you can integrate libraries seamlessly into existing projects to add functionality. Incorporating frameworks seamlessly into an existing project is impossible. Instead, frameworks should be used when starting a new project.

3. What is CDN? Why to use CDN?
Ans:-
A content delivery network (CDN) is a group of servers spread out over a region or around the world that work together to speed up content delivery on the web. They offer four important benefits: better performance i.e. 50% reductions in load times, increased reliability, cost savings, and resilience against cyber attacks.

4. Why is React known as React? 
Ans:- 
React is named React because of its ability to react to changes in data. When the data in a React component changes, React will automatically re-render the component so that it reflects the new data. This makes it easy to create performant user interfaces that always look up-to-date.

5.What is cross-origin in script tag?
Ans:- It provides support for CORS, defining how the element handles cross-origin requests(data fetch from different ports), thereby enabling the configuration of the CORS requests for the element's fetched data. Depending on the element, the attribute can be a CORS settings attribute.

6. Difference between React and ReactDOM
Ans:-
React library is responsible for creating views and ReactDOM library is responsible to actually render UI in the browser with .render() . The reason React and ReactDOM were split into two libraries was due to the arrival of React Native. React contains functionality utilized in web and mobile apps. ReactDOM functionality is utilized only in web apps.

7. What is difference between react.development.js and react.production.js files via CDN?
Ans:-
The versions react.development.js  are only meant for development, and are not suitable for production. Minified and optimized production versions of React are available at react.production.js

8. What is async and defer? (taking reference from Akshay's video)
Ans:- 

Async and Defer are Boolean attributes which are used along with script tag to load external scripts efficiently into our web page.
If you load a webpage, there are 2 things happening in browser: -
1.HTML parsing by browser
2.Loading of the scripts. It contains 2 parts as below: -
1.	Fetching scripts from Network 
2.	Executing scripts line by line
In normal script tag, JavaScript script tag blocks the parsing of HTML by browser in between as the script tag fetched and executes then and there. After that only parsing continues.
With Async attribute, any script with this attribute , scripts are fetched from the network asynchronously along with HTML parsing. As soon as scripts are fetched and available in browser,the HTML parsing stops and the scripts are executed then and there. And once the scripts are executed line by line,HTML parsing continues. 

With defer attribute, HTML parsing goes on and the scripts are fetched in parallel. And then when HTML parsing completes, then only these scripts get executed at the end.
So, to conclude, Async attribute doesn’t guarantee the order of execution of multiple dependent scripts. 
Defer maintains the order of execution of scripts when the HTML parsing is completed.
Use Async when the scripts are independent of each other else use Defer